laying out a major refactor of the plugin
Taking us from a very script-ish form to a more extensible design that
will allow a programmer to extend the plugin without modification to
the plugin core. We also want to remove any domain classes and
controllers to allow the application developer the freedom to create
their own in any format that they want.